[OpenMP][MLIR][Flang] Add specialised offloading lowering of fortran allocatable types to OpenMPToLLVMIRTranslation
This patch seeks to add a specialsied lowering of mapping
of Fortran allocatables/pointer types to the OpenMP -> LLVM-IR
lowering phase of the OpenMP dialect.

These types come wrapped in a structure (descriptor) which
currently must be mapped alongside the data (a pointer inside
of the descriptor structure) they hold. This descriptor structure
contains information such as type size of elements, upper bounds,
lower bounds, and other information. The current target lowering
will lower the region with the expectation that this descriptor is
mapped across to the kernel and useable inside of the kernel to
perform various instructions. I believe future OpenMP wording
is also leaning towards explicit mapping of descriptors for Fortran
allocatables in any case.

So we must lower both the descriptor and the internal pointer that
points to the user specified data that they wish to access on device.

This patch does so utilising a similar mapping to how Clang maps
structures containing pointer members, e..g.

map(tofrom: descriptor, descriptor->data)

Where the descriptors data is mapped as a member_of the larger
structure. The offset and size to be offloaded are currently
calculated by the data stored in the descriptor.