[MLIR][OpenMP] Moved descriptor base_addr from varPtr to varPtrPtr field
This required making the varPtr optional alongside varPtrPtr, and adding verification to test at least one of these was specified.

I also added the varPtrPtrType field and renamed VarType to varPtrType, so
that both fields have a distinct type (verification was also used to make sure
the respective type field was provided). This change is not completely neccesary it is likely they could share the same type, but this wouldn't
allow for both a var_ptr and var_ptr_ptr to be specified on the one
map_info operation reasonably (which is allowed currently).

The lowering and tests were also updated to reflect these changes.